fix:controller분리,userName athentication사용
