+++ src/main/java/com/hwachang/hwachangapi/domain/consultingRoomModule/controller/CategoryController.java
@@ -1,0 +1,21 @@
+package com.hwachang.hwachangapi.domain.consultingRoomModule.controller;
+
+import com.hwachang.hwachangapi.domain.consultingRoomModule.dto.CategoryDto;
+import com.hwachang.hwachangapi.domain.consultingRoomModule.service.CategoryService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import java.util.List;
+
+@RestController
+@RequestMapping("/category")
+@RequiredArgsConstructor
+public class CategoryController {
+    private final CategoryService categoryService;
+
+    @GetMapping()
+    public List<CategoryDto> getCategories() {
+        return this.categoryService.getCategories();
+    }
+}
